+++ python/llm/example/CPU/HF-Transformers-AutoModels/Streaming/streaming_llm/enable_streaming_llm.py
@@ -47,7 +47,7 @@ from .kv_cache import StartRecentKVCache
 def enable_streaming_llm(model, start_size, recent_size):
     if "llama" in model.config.model_type:
         k_seq_dim = v_seq_dim = 2
-        from .pos_shift.modify_llama import (
+        from .modify_llama import (
             enable_llama_pos_shift_attention,
         )
 
@@ -57,7 +57,7 @@ def enable_streaming_llm(model, start_size, recent_size):
         k_seq_dim = 3
     elif "gpt_neox" in model.config.model_type:
         k_seq_dim = v_seq_dim = 2
-        from .pos_shift.modify_gpt_neox import (
+        from .modify_gpt_neox import (
             enable_gpt_neox_pos_shift_attention,
         )
 
@@ -65,7 +65,7 @@ def enable_streaming_llm(model, start_size, recent_size):
     elif "falcon" in model.config.model_type:
         v_seq_dim = 1
         k_seq_dim = 1
-        from .pos_shift.modify_falcon import (
+        from .modify_falcon import (
             enable_falcon_pos_shift_attention,
         )
