+++ python/llm/example/CPU/HF-Transformers-AutoModels/Streaming/streaming_llm/modify_gpt_neox.py
